Move Fabric initializers to Fabric project
