+++ FabricMod/src/main/kotlin/opekope2/avm_staff/internal/fabric/StaffMod.kt
@@ -16,7 +16,7 @@
  * along with this mod. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package opekope2.avm_staff.internal
+package opekope2.avm_staff.internal.fabric
 
 import net.fabricmc.api.ModInitializer
 import net.fabricmc.fabric.api.event.player.AttackBlockCallback
@@ -36,20 +36,19 @@ import net.minecraft.util.hit.EntityHitResult
 import net.minecraft.util.math.BlockPos
 import net.minecraft.util.math.Direction
 import net.minecraft.world.World
+import opekope2.avm_staff.IStaffMod
 import opekope2.avm_staff.api.item.StaffItem
 import opekope2.avm_staff.internal.packet.c2s.play.AddItemToStaffC2SPacket
 import opekope2.avm_staff.internal.packet.c2s.play.RemoveItemFromStaffC2SPacket
 import opekope2.avm_staff.internal.packet.c2s.play.StaffAttackC2SPacket
 import opekope2.avm_staff.internal.server.StaffPacketHandler
+import opekope2.avm_staff.util.MOD_ID
 import opekope2.avm_staff.util.handlerOfItem
 import opekope2.avm_staff.util.itemInStaff
 
 @Suppress("unused")
-object StaffMod : ModInitializer {
-    const val MOD_ID = "avm_staff"
-
-    @JvmField
-    val STAFF_ITEM: StaffItem = Registry.register(
+object StaffMod : ModInitializer, IStaffMod {
+    override val staffItem: StaffItem = Registry.register(
         Registries.ITEM,
         Identifier(MOD_ID, "staff"),
         StaffItem(FabricItemSettings().maxCount(1))
@@ -57,10 +56,10 @@ object StaffMod : ModInitializer {
 
     override fun onInitialize() {
         ItemGroupEvents.modifyEntriesEvent(ItemGroups.TOOLS).register { entries ->
-            entries.addAfter(Items.NETHERITE_HOE, STAFF_ITEM)
+            entries.addAfter(Items.NETHERITE_HOE, staffItem)
         }
         ItemGroupEvents.modifyEntriesEvent(ItemGroups.COMBAT).register { entries ->
-            entries.addAfter(Items.TRIDENT, STAFF_ITEM)
+            entries.addAfter(Items.TRIDENT, staffItem)
         }
 
         AddItemToStaffC2SPacket.registerGlobalReceiver(StaffPacketHandler::addBlockToStaff)
@@ -79,7 +78,7 @@ object StaffMod : ModInitializer {
         direction: Direction
     ): ActionResult {
         val staffStack = player.getStackInHand(hand)
-        if (!staffStack.isOf(STAFF_ITEM)) return ActionResult.PASS
+        if (!staffStack.isOf(staffItem)) return ActionResult.PASS
 
         return staffStack.itemInStaff?.handlerOfItem?.attackBlock(staffStack, world, player, target, direction, hand)
             ?: ActionResult.PASS
@@ -95,7 +94,7 @@ object StaffMod : ModInitializer {
         if (world.isClient) return ActionResult.PASS // Handled with Staff Mod mixin
 
         val staffStack = player.getStackInHand(hand)
-        if (!staffStack.isOf(STAFF_ITEM)) return ActionResult.PASS
+        if (!staffStack.isOf(staffItem)) return ActionResult.PASS
 
         return staffStack.itemInStaff?.handlerOfItem?.attackEntity(staffStack, world, player, target, hand)
             ?: ActionResult.PASS
