+++ FabricMod/src/main/kotlin/opekope2/avm_staff/internal/fabric/StaffModClient.kt
@@ -16,7 +16,7 @@
  * along with this mod. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package opekope2.avm_staff.internal.client
+package opekope2.avm_staff.internal.fabric
 
 import net.fabricmc.api.ClientModInitializer
 import net.fabricmc.api.EnvType
@@ -29,7 +29,7 @@ import net.minecraft.client.MinecraftClient
 import net.minecraft.client.option.KeyBinding
 import net.minecraft.client.render.model.BakedModel
 import net.minecraft.client.util.InputUtil
-import opekope2.avm_staff.internal.StaffMod.MOD_ID
+import opekope2.avm_staff.util.MOD_ID
 import opekope2.avm_staff.internal.model.StaffItemModel
 import opekope2.avm_staff.internal.packet.c2s.play.AddItemToStaffC2SPacket
 import opekope2.avm_staff.internal.packet.c2s.play.RemoveItemFromStaffC2SPacket
